+++ src/components/phone/module.ts
@@ -1,0 +1,26 @@
+import {NgModule} from '@angular/core';
+import {CommonModule} from '@angular/common';
+import {NG_VALUE_ACCESSOR, ReactiveFormsModule} from '@angular/forms';
+import {CsIconModule} from '../icon/module';
+
+import {PhoneLocalization} from '../../models/phone_number';
+
+import {CsPhoneInput, PhoneInputControlValueAccessor} from './phone-input';
+import {PhonePipe} from './phone-pipe';
+
+export {CsPhoneInput, PhonePipe};
+
+@NgModule({
+    imports: [CommonModule, CsIconModule, ReactiveFormsModule],
+    declarations: [
+        CsPhoneInput,
+    ],
+    providers: [
+        PhoneLocalization,
+        {provide: NG_VALUE_ACCESSOR, useClass: PhoneInputControlValueAccessor, multi: true}
+    ],
+    exports: [
+        CsPhoneInput,
+    ]
+})
+export class PhoneModule {}